Merged PR 18: PSTO Staff should be able to Create SetUpCustomer
Our current version is only allowing to create SetUpCustomer and SetUpCustomerBusiness as  as request but we should be able to let the PSTO Staff create old SetUpCustomers so for this function:
An add button should appear in the SetUpCustomer page (will be created by Task 50).
When an add button is pressed a modal will appear just like the Customer Application.
PSTO Staff will fill out all the information when the user press save:
1. This will create a new SetUpCustomer and SetUpCustomerBusiness Entry
2. Will create a new user
3. That user will also receive and email for the their account details but this time the manifestation of intent will not be inlcuded